% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC.r
\name{ABCLogPosterior}
\alias{ABCLogPosterior}
\title{ABC logged posterior distribution}
\usage{
ABCLogPosterior(epsilon, sum.stats, distanceABC, fitmodel, theta, init.state,
  data)
}
\arguments{
\item{epsilon}{numeric vector, ABC tolerances for distances between data and simulations. If a vector of length 1 and the distance function returns a vector of distances, this will be expanded to be same tolerance for all the parameters.}

\item{sum.stats}{a list of functions to calculate summary statistics. Each of these takes one argument (a trajectory with an "obs" column) and returns a number (the summary statistic given the trajectory)}

\item{distanceABC}{a function that take three arguments: \code{sum.stats}, a list of summary statistics, \code{data.obs} (the data trajectory of observations) and \code{model.obs} (a model trajectory of observations), and returns the distance between the model run and the data in terms of the summary statistics}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{init.state}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{"time"} and the observation column must be named \code{"obs"}.}
}
\value{
a list of two elements
\itemize{
     \item \code{log.density} numeric, logged value of the ABC posterior distribution evaluated at \code{theta}
     \item \code{trace} named vector with trace information (theta, distance, log.density)
}
}
\description{
This function evaluates the ABC posterior distribution at \code{theta} (using a single simulation trajectory) and returns the result in a suitable format for \code{\link{mcmcMH}}.
}
\seealso{
computeDistanceABC
}
