% Generated by roxygen2 (4.0.1): do not edit by hand
\name{agcfit}
\alias{agcfit}
\title{Wrapper function for automated growth curve fitting}
\usage{
agcfit(data, w_size, od_name, time_name, pearCutoff = 0.95, method = "lsw",
  parallel = FALSE, progress = "text")
}
\arguments{
\item{data}{long-form data frame with growth data}

\item{w_size}{size of sliding window (number of datapoints)}

\item{od_name}{name of column containing OD values}

\item{time_name}{name of column containing times (in units after start of experiment)}

\item{pearCutoff}{Only fits with a correlation coefficient higher than the cutoff are considered. Defaults to 0.95.}

\item{method}{the method used to fit, at the moment only "lsw" for a linear sliding window fit.}

\item{parallel}{if \code{TRUE}, apply function in parallel, using parallel backend provided by \code{\link[foreach]{foreach}}.}

\item{progress}{name of the progress bar to use, see \code{\link[plyr]{create_progress_bar}}}
}
\description{
Wrapper function for automated growth curve fitting
}
\section{Output}{

   none
}
\keyword{growthcurve}

