% Generated by roxygen2 (4.0.1): do not edit by hand
\name{burnAndThin}
\alias{burnAndThin}
\title{Burn and thin MCMC chain}
\usage{
burnAndThin(trace, burn = 0, thin = 0)
}
\arguments{
\item{trace}{a \code{data.frame} with all variables in column, as outputed by \code{\link{mcmcMH}}.}

\item{burn}{proportion of the chain to burn.}

\item{thin}{number of samples to discard per sample that is being kept}
}
\value{
a \code{data.frame}.
}
\description{
Return a burned and thined trace of the chain.
}

