% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{calibrateSmc}
\alias{calibrateSmc}
\title{A data set resulting from calibrating the number of particles in a
particleFilter to estimate the likelihood of the stochastic SEIT4L
model to the influenza outbreak on Tristan da Cunha}
\format{
A data frame with 11 rows and 5 columns describing the results of the
  calibration procedure
}
\description{
A dataset containing the results of an attempt to calibrate the number of
particles to be used in estimating the likelihood when fitting the
`seit4lStoch` model to the `fluTdc1971` dataset. Estimates were obtained
using 100 replicates of the estimation procedure.
}
\details{
\itemize{
  \item \code{nParticles} number of particles
  \item \code{mean} mean estimated likelihood
  \item \code{sd} standard deviatoin of the estimated likelihood
  \item \code{propDepleted} proportion of particles depleted
  \item \code{days} number of days the 100 replicated took to estimate
}
}
