% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikelihood.r
\name{computeDic}
\alias{computeDic}
\title{Compute the DIC}
\usage{
computeDic(trace, fitmodel, initState, data, margLogLike = dTrajObs, ...)
}
\arguments{
\item{trace}{either a \code{data.frame} or \code{mcmc} object. Must contain
one column with the posterior \code{logLikelihood}.}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{initState}{named numeric vector. Initial values of the state
variables. Names should match \code{fitmodel$stateNames}.}

\item{data}{data frame. Observation times and observed data. The time column
must be named \code{"time"} and the observation column must be named
\code{"obs"}.}

\item{margLogLike}{\R-function to compute the marginal log-likelihood of
\code{theta}.}

\item{...}{further arguments to be passed to \code{margLogLike}}
}
\value{
a list of 5 elements:
\itemize{
    \item \code{DIC} value of the DIC
    \item \code{thetaBar} mean posterior of theta
    \item \code{logLikeThetaBar} log-likelihood of \code{thetaBar}
    \item \code{DThetaBar} deviance of \code{thetaBar}
    \item \code{pD} effective number of parameters
}
}
\description{
This function computes the Deviance Information Criterion (DIC) of a
\code{\link{fitmodel}} from a MCMC sample.
}
