% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikelihood.r
\name{dLogPosteriorWrapper}
\alias{dLogPosteriorWrapper}
\title{A wrapper for \code{dLogPosterior}}
\usage{
dLogPosteriorWrapper(fitmodel, initState, data, margLogLike, ...)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{initState}{named numeric vector. Initial values of the state
variables. Names should match \code{fitmodel$stateNames}.}

\item{data}{data frame. Observation times and observed data. The time column
must be named \code{"time"} and the observation column must be named
\code{"obs"}.}

\item{margLogLike}{\R-function to compute the marginal log-likelihood of
\code{theta}.}

\item{...}{further arguments to be passed to \code{margLogLike}}
}
\value{
a \R-function with one argument called \code{theta}.
}
\description{
A wrapper for \code{\link{dLogPosterior}} that returns a function that can be
used as a \code{target} for \code{\link{mcmcMh}}
}
\keyword{internal}
