% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.attrErrorCodes}
\alias{.attrErrorCodes}
\title{add attributes with error codes to data.frame}
\usage{
.attrErrorCodes(best, min_numP, RsqCutoff, growthCheck)
}
\arguments{
\item{best}{a data.frame from \code{\link{"pickfit"}}}
}
\description{
add attributes with error codes to data.frame
}
\section{Output}{

   data.frame with attribute \code{"error codes"}
}

\keyword{fitr}
