% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.gcDataTrafo}
\alias{.gcDataTrafo}
\title{transform OD values}
\usage{
.gcDataTrafo(data, od_colnr, time_colnr, trafo, logBase = 2)
}
\arguments{
\item{data}{data}

\item{od_colnr}{col with OD values}

\item{time_colnr}{col with times}

\item{trafo}{which trafo to use}

\item{logBase}{which logbase (default: 2)}
}
\description{
internal function to transform od values. -Inf, NaN values are replaced with NA.
}
\section{Output}{

   data.frame as input with new colum 'ODtrans'
}

\keyword{fitr}
