% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accFunctions.R
\name{.rgbColorGradient}
\alias{.rgbColorGradient}
\title{generate color gradient}
\usage{
.rgbColorGradient(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
with colors from blue(low) to red(high)
NA get the same color as the lowest value.
}
\section{Output}{

   character vector with colors
}

\keyword{fitr}
