% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{epi3}
\alias{epi3}
\title{Time-series of an outbreak simulated using an SIR model.}
\format{
A data frame with 44 rows and 2 variables
}
\description{
A dataset containing the daily observed incidence of new infections for a
simulated outbreak from the `sirDeter` model with R_0 = 2.1 and D_inf = 3, in
a population of 999 susceptibles and a single infected person.
}
\details{
\itemize{
  \item \code{time} time step
  \item \code{obs} number of observations
}
}
\keyword{internal}
