% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{epi4}
\alias{epi4}
\title{Time-series of an outbreak simulated using an SIR model with underreporting.}
\format{
A data frame with 100 rows and 2 variables
}
\description{
A dataset containing the daily observed incidence of new infections for a
simulated outbreak from the `sirReportingDeter` model with R_0 = 2.4, D_inf =
8 and RR = 0.25, in a population of 999 susceptibles and a single infected
person.
}
\details{
\itemize{
  \item \code{time} time step
  \item \code{obs} number of observations
}
}
