% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-class.R
\docType{class}
\name{fitR-class}
\alias{fitR-class}
\title{A fitR object}
\description{
Object of class \code{fitR} are created by the \code{fitR} function
}
\details{
An object of the class `fitR' has the following slots:
\itemize{
\item \code{statistics} A matrix with one row for each competing model, and one column
for each fit statistic calculated.
}
}
\author{
David R. Miller: \email{drmiller@wustl.edu}
}
