\name{FitR-package}
\alias{FitR-package}
\alias{FitR}
\docType{package}
\title{
\packageTitle{FitR}
}
\description{
\packageDescription{FitR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FitR}
\packageIndices{FitR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FitR}

Maintainer: \packageMaintainer{FitR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
