% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRstats.R
\docType{methods}
\name{fitRstats}
\alias{fitRstats}
\alias{fitRstats,ANY-method}
\title{fitRstats}
\usage{
fitRstats(true, predicted, naive, statistics, ...)

\S4method{fitRstats}{ANY}(true, predicted, naive = NULL,
  statistics = "rmse", ...)
}
\arguments{
\item{true}{A vector of the true values}

\item{predicted}{A matrix of values predicted by competing models, with one column for
each model}

\item{naive}{A vector of naive forecasts}

\item{statistics}{A vector of character strings indicating the fit statistics to
be calculated}

\item{...}{Other arguments to be passed through}
}
\value{
A matrix with one row for each competing model, and one column for each fit 
statistic calculated
}
\description{
Calculates the fit statistics for competing models by comparing predictive accuracy
}
\author{
David R. Miller
}
