% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitstat.R
\name{fitstats}
\alias{fitstats}
\title{Calculate Fit Statistics}
\usage{
fitstats(y, P, r = NULL, HideFits = c("MRAE"))
}
\arguments{
\item{y}{A vector of test data values}

\item{P}{A matrix of predicted values}

\item{r}{A vector of naive estimates}

\item{HideFits}{A vector of fit statistics the user wants suppressed from the output. abbreviations as strings. 
Possible values include RMSE, MAD, RMSLE, MAPE, MEAPE, and MRAE (hidden by default since 
it is for use with the r argument only)}
}
\value{
A matrix of fit statistic values, where rows are models and columns 
are fit statistics
}
\description{
Finds fit statistics given a matrix of predicted values and a vector
of test data. Providing a vector of naive predictions allows for the inclusion
of MRAE in fit statistic calculations. Note that NA values are case-wise deleted.
}
\author{
Ryden W. Butler
}
