% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lFitFunctions.R
\name{flufit}
\alias{flufit}
\title{Estimate slope of fluorescence data}
\usage{
flufit(data, w_size, flu_name, time_name, logBase = 2)
}
\arguments{
\item{data}{long-form data frame with growth data.}

\item{w_size}{size of sliding window (number of datapoints).}

\item{flu_name}{name of column containing fluorescence values.}

\item{time_name}{name of column containing times (in units after start of experiment).}

\item{logBase}{Base for logarithmitic transformation if trafo != "none", defaults to 2.}
}
\description{
fits a a linear model over a sliding window to fluorescence data. Outputs all fits as a \code{data.frame}
}
\section{Output}{

   \code{data.frame} of with all fits for all possible windows
}
\keyword{fitr,}
\keyword{growthcurve}

