% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{FluTdC1971}
\alias{FluTdC1971}
\title{Time-series of the 1971 influenza epidemic in Tristan-da-Cunha}
\format{A data frame with 59 rows and 3 variables}
\source{
\url{http://eutils.ncbi.nlm.nih.gov/entrez/eutils/elink.fcgi?dbfrom=pubmed&id=4511951&retmode=ref&cmd=prlinks}
}
\description{
A dataset containing the daily incidence recorded during the 1971 influenza A/H3N2 two-wave epidemic on the island of Tristan-da-Cunha.
}
\details{
\itemize{
  \item \code{date} calendar date of the record 
  \item \code{time} day of record since beginning of epidemic
  \item \code{Inc} daily count incidence of influenza-like-illness
}
}

