% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gcblanking}
\alias{gcblanking}
\title{Function for blanking OD data & flagging growth/no growth}
\usage{
gcblanking(data, group, od_name, bl_col, bl_name, cutoff = 0)
}
\arguments{
\item{data}{long-form data frame with growth data, must contain column named ID with a unique ID for each curve.}

\item{group}{name of grouping variable (e.g. plates for which blanks must be seperate), a factor.}

\item{od_name}{name of column containing OD values}

\item{bl_col}{name of column containing well description, e.g. "strain"}

\item{bl_name}{which wells are blanks?, e.g. "bl"}

\item{cutoff}{discard values <= cutoff}
}
\description{
Function for blanking OD data & flagging growth/no growth
}
\section{Output}{

   long-form data frame with additional columns od_name_bl (blanked OD values) and QCgrowth (logical)
}
\keyword{blanking}
\keyword{fitr,}
\keyword{growthcurve,}

