% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gcfit}
\alias{gcfit}
\title{fit growth curve}
\usage{
gcfit(data, w_size, od_name, time_name, trafo = "logNN0")
}
\arguments{
\item{data}{long-form data frame with growth data}

\item{w_size}{size of sliding window (number of datapoints)}

\item{od_name}{name of column containing OD values}

\item{time_name}{name of column containing times (in units after start of experiment)}

\item{trafo}{Data transformation, one of "logNN0" (for log(N/N0)), "log" or "none".}
}
\description{
fits a model to growth data. Outputs all fits as a \code{data.frame}
}
\section{Output}{

   \code{data.frame} of with all fits for all possible windows
}
\keyword{growthcurve}

