% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genObsTraj}
\alias{genObsTraj}
\title{Generate an observation trajectory for a fitmodel}
\usage{
genObsTraj(fitmodel, theta, state.init, times)
}
\arguments{
\item{times}{the times at which to generate observations}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{state.init}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}
}
\value{
numeric value of the log-likelihood
}
\description{
This function simulates a model defined in a \code{\link{fitmodel}}
object and generates observations at each time point. It returns
the trajectory with an additions \code{obs} column.
}
\seealso{
\code{\link{trajLogLike}}
}

