% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitr.r
\docType{data}
\name{growthDataFive}
\alias{growthDataFive}
\title{Five growth curve}
\format{A data frame with 485 rows and 10 variables}
\usage{
data(growthDataFive)
}
\description{
A dataset containing five growth curve
'
\itemize{
  \item ID. unique ID of growth curve
  \item Row. Row position on 96 well plate
  \item Col. Column position on 96 well plate
  \item Strain. bacterial strain descriptor
  \item Rep. Replicate
  \item ara. conc. of Arabinos
  \item tet. conc. of tetracycline
  \item OD. Optical density measurement
  \item time. time of measurement (min. after start of experiment)
  \item ODred. blank corrected OD measurements
}
}
\keyword{datasets,}
\keyword{fitr}

