% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitr.r
\docType{data}
\name{growthDataOne}
\alias{growthDataOne}
\title{One growth curve}
\format{
A data frame with 97 rows and 10 variables
}
\usage{
data(growthDataOne)
}
\description{
A dataset containing one growth curve
'
\itemize{
  \item ID. unique ID of growth curve  
  \item Row. Row position on 96 well plate
  \item Col. Column position on 96 well plate 
  \item Strain. bacterial strain descriptor
  \item Rep. Replicate
  \item ara. conc. of Arabinos
  \item tet. conc. of tetracycline
  \item OD. Optical density measurement
  \item time. time of measurement (min. after start of experiment)
  \item ODred. blank corrected OD measurements
}
}
\keyword{datasets,}
\keyword{fitr}
