% Generated by roxygen2 (4.0.1): do not edit by hand
\name{linfit}
\alias{linfit}
\title{fit growth curve using sliding linear fits}
\usage{
linfit(data, w_size, od_name, time_name)
}
\arguments{
\item{data}{long-form data frame with growth data}

\item{w_size}{size of sliding window (number of datapoints)}

\item{od_name}{name of column containing OD values}

\item{time_name}{name of column containing times (in units after start of experiment)}
}
\description{
fits linear fits in a sliding window to growth data. Outputs all fits as a \code{data.frame}
}
\section{Output}{

   \code{data.frame} of with all fits for all possible windows
}
\keyword{growthcurve}

