% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLikelihood.r
\name{logPosteriorWrapper}
\alias{logPosteriorWrapper}
\title{A wrapper for \code{logPosterior}}
\usage{
logPosteriorWrapper(fitmodel, init.state, data, margLogLike, ...)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{init.state}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{"time"} and the observation column must be named \code{"obs"}.}

\item{margLogLike}{\R-function to compute the marginal log-likelihood of \code{theta}.}

\item{...}{further arguments to be passed to \code{margLogLike}}
}
\value{
a \R-function with one argument called \code{theta}.
}
\description{
A wrapper for \code{\link{logPosterior}} that returns a function that can be used as a \code{target} for \code{\link{mcmcMH}}
}

