% Generated by roxygen2 (4.0.1): do not edit by hand
\name{margLogLikeDeter}
\alias{margLogLikeDeter}
\title{Marginal log-likelihood for a deterministic model}
\usage{
margLogLikeDeter(fitmodel, theta, state.init, data)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{state.init}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{time}, whereas the name of the data column should match the one used in the function \code{fitmodel$logLikePoint}.}
}
\value{
numeric value of the log-likelihood
}
\description{
Compute the marginal log-likelihood of \code{theta} for a deterministic model defined in a \code{\link{fitmodel}} object.
}

