% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{mcmcEpi4}
\alias{mcmcEpi4}
\title{Markov-chain Monte Carlo (MCMC) trace of a fit of the SIR model to a
simulated outbreak}
\format{
A data frame with 10001 rows and 2 variables
}
\description{
A dataset containing the MCMC trace of a fit of the `sirDeter` model to the
`epi4` data set.
}
\details{
\itemize{
  \item \code{R_0} basic reproduction number
  \item \code{D_inf} duration of infection
  \item \code{RR} reporting rate
}
}
