% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{mcmcSeit4lInfoPriorTheta2}
\alias{mcmcSeit4lInfoPriorTheta2}
\title{Markov-chain Monte Carlo (MCMC) outputs of a fit of the SEIT4L model to the
influenza outbreak on Tristan da Cunha (long run)}
\format{
A list with three elements describing the MCMC outputs
}
\description{
A dataset containing the MCMC outputs of a fit of the `seit4lDeter` model to
the `fluTdc1971` data set with informative priors (50000 iterations). This
differs from the `mcmcSeit4lTheta2` dataset in the choice of priors and from
the `mcmcSeit4lInfoPriorTheta1` dataset in the choice of initial parameter
set theta.
}
\details{
\itemize{
  \item \code{trace} data frame containing the MCMC trace
  \item \code{acceptanceRate} a single value denoting the acceptance rate
  \item \code{covmatEmprical} the covariance matrix of parameter samples
}
}
\keyword{internal}
