% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{mcmcSeitl}
\alias{mcmcSeitl}
\title{Markov-chain Monte Carlo (MCMC) outputs of a fit of the SEITL model to the
influenza outbreak on Tristan da Cunha (short run)}
\format{
A list with three elements describing the MCMC outputs
}
\description{
A dataset containing the MCMC outputs of a fit of the `seitlDeter` model to
the `fluTdc1971` data set (5000 iterations).
}
\details{
\itemize{
  \item \code{trace} data frame containing the MCMC trace
  \item \code{acceptanceRate} a single value denoting the acceptance rate
  \item \code{covmatEmprical} the covariance matrix of parameter samples
}
}
\keyword{internal}
