% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pickfit}
\alias{pickfit}
\title{pick best fits}
\usage{
pickfit(fits, pearCutoff = 0.95)
}
\arguments{
\item{fits}{\code{data.frame} containing fits generated by \code{\link{lfit}}}

\item{pearCutoff}{Only fits with a correlation coefficient higher than the cutoff are considered. Defaults to 0.95.}
}
\description{
pick the 'most valid' fit. This will be the fit with the highest mumax and a correlation coefficient above \code{pearCoeff}
}
\section{Output}{

   \code{data.frame} with all fits from lfit
}
\keyword{fitr,}
\keyword{growthcurve}

