% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.fitr}
\alias{plot.fitr}
\title{plot function for class "fitr"}
\usage{
\method{plot}{fitr}(fitr, interactive = TRUE, select = FALSE,
  sample_size = 5)
}
\arguments{
\item{object}{of class "fitr"}
}
\description{
plots data and best fits, a wrapper for \code{\link{plot_fitr}}
}
\keyword{fitr}

