% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lFitFunctions.R
\name{plot_fitr}
\alias{plot_fitr}
\title{plot fits generated with gcfit and selected with pickfit}
\usage{
plot_fitr(bestfit, fits, data, od_name, time_name, interactive = TRUE,
  select = FALSE, sample_size = 5, save = FALSE)
}
\arguments{
\item{bestfit}{\code{data.frame} with best fit generated by \code{\link{pickfit}}}

\item{fits}{\code{data.frame} containing fits generated by \code{\link{gcfit}}}

\item{data}{long-form data frame with growth data}

\item{od_name}{name of column containing OD values}

\item{time_name}{name of column containing times (in units after start of experiment)}

\item{interactive}{boolean; Should every plot be shown? Advance to next plot by clicking in plot area.}

\item{select}{Either a vector containing IDs of a selection of fits to plot (IDs not found in \code{data} are quietly ignored), or one of \code{c("sample","sampleQ","failed")}. \code{"sample"} will draw an uniform sample of size \code{sample_size}. \code{"sampleQ"} will draw a total (rounded) of \code{sample_size} samples split over growthrate (mumax) sample quantiles. \code{"failed"} will display fits were no best fit could be selected.  Defaults to \code{FALSE}, showing all fits.}

\item{sample_size}{Size of sample to draw, defaults to the arbitrary number of 5}
}
\description{
plot all fits, or only a selection using the \code{select} argument.
}
\section{Output}{

   none
}
\keyword{fitr,}
\keyword{growthcurve}

