% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotESSBurn}
\alias{plotESSBurn}
\title{Plot Effective Sample Size (ESS) against burn-in}
\usage{
plotESSBurn(trace, longest.burn.in = ifelse(length(trace) > 1,
  nrow(trace[[1]]), nrow(trace))/2, step.size = round(longest.burn.in/50))
}
\arguments{
\item{trace}{A data frame of an MCMC chain with one column per parameter}

\item{longest.burn.in}{The longest burn in to test. Defaults to half the length of the trace}

\item{step.size}{The size of the steps of burn-in to test. Defaults to 1/50th of \code{longest.burn.in}}
}
\value{
a plot of the ESS against burn.in
}
\description{
Takes an mcmc trace and tests the ESS at different values of burn-in
}

