% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotFit}
\alias{plotFit}
\title{Plot fit of model to data}
\usage{
plotFit(fitmodel, theta, init.state, data, n.replicates = 1, summary = TRUE,
  alpha = min(1, 10/n.replicates), all.vars = FALSE, p.extinction = FALSE,
  plot = TRUE)
}
\arguments{
\item{n.replicates}{numeric, number of replicated simulations.}

\item{all.vars}{logical, if \code{FALSE} only the observations are plotted. Otherwise, all state variables are plotted.}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{init.state}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{"time"} and the observation column must be named \code{"obs"}.}

\item{summary}{logical. If \code{TRUE}, the mean, median as well as the 50th and 95th percentile of the trajectories are plotted (default). If \code{FALSE}, all individual trajectories are plotted (transparency can be set with \code{alpha}).}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{p.extinction}{logical. If \code{TRUE}, the time-series of the proportion of faded-out epidemics is plotted (default to \code{FALSE}). This is only relevant for stochastic models.}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\value{
if \code{plot==FALSE}, a list of 2 elements is returned:
\itemize{
    \item \code{simulations} \code{data.frame} of \code{n.replicates} simulated observations.
    \item \code{plot} the plot of the fit.
}
}
\description{
This function simulates the model under \code{theta}, generates observation and plot them against the data. Since simulation and observation processes can be stochastic, \code{n.replicates} can be plotted.
}

