% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotHPDregion2D}
\alias{plotHPDregion2D}
\title{2D highest posterior density region}
\usage{
plotHPDregion2D(
  trace,
  vars,
  prob = c(0.95, 0.75, 0.5, 0.25, 0.1),
  xlab = NULL,
  ylab = NULL,
  plot = TRUE
)
}
\arguments{
\item{trace}{either a \code{data.frame} or \code{mcmc} object.}

\item{vars}{which variables to plot: numeric or character vector}

\item{prob}{probability level}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\description{
Given a sample from a multivariate posterior distribution, plot the bivariate
region of highest marginal posterior density (HPD) for two variables with
defined levels.
}
\note{
HPD levels are computed using the function
  \code{\link[emdbook]{HPDregionplot}} from the package \code{emdbook}.
}
