% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotPosteriorDensity}
\alias{plotPosteriorDensity}
\title{Plot MCMC posterior densities}
\usage{
plotPosteriorDensity(trace, prior = NULL, colour = NULL, plot = TRUE)
}
\arguments{
\item{trace}{either a \code{data.frame} or a \code{list} of \code{data.frame}
with all variables in column, as outputed by \code{\link{mcmcMh}}. Accept
also an \code{mcmc}, a \code{mcmc.list} object or a \code{list} of
\code{mcmc.list} .}

\item{prior}{a \code{data.frame} containing the prior density. It must have
  the three following columns:
\itemize{
    \item \code{theta} names of the parameters
    \item \code{x} value of the parameters
    \item \code{density} density of the prior at \code{x}
}}

\item{colour}{named vector of two characters and containing colour names for
posterior and prior distributions. Vector names must be \code{posterior}
and \code{prior}.}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\description{
Plot the posterior density.
}
\examples{
data(mcmcEpi)
plotPosteriorDensity(mcmcEpi1)
}
\seealso{
burnAndThin
}
