% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSMC}
\alias{plotSMC}
\title{Plot result of SMC}
\usage{
plotSMC(smc, fitmodel, theta, data = NULL, summary = TRUE, alpha = 1,
  all.vars = FALSE, plot = TRUE)
}
\arguments{
\item{smc}{output of \code{\link{particleFilter}}}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{time}, whereas the name of the data column should match one of \code{fitmodel$state.names}.}

\item{summary}{logical. If \code{TRUE}, the mean, median as well as the 50th and 95th percentile of the trajectories are plotted (default). If \code{FALSE}, all individual trajectories are plotted (transparency can be set with \code{alpha}).}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{all.vars}{logical, if \code{FALSE} only the observations are plotted. Otherwise, all state variables are plotted.}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\description{
Plot the observation generated by the filtered trajectories together with the data.
}
\seealso{
particleFilter
}

