% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotTrace}
\alias{plotTrace}
\title{Plot MCMC trace}
\usage{
plotTrace(trace, estimatedOnly = FALSE)
}
\arguments{
\item{trace}{a \code{data.frame} with one column per estimated parameter, as
returned by \code{\link{burnAndThin}}}

\item{estimatedOnly}{logical, if \code{TRUE} only estimated parameters are
displayed.}
}
\description{
Plot the traces of all estimated variables.
}
\examples{
data(mcmc.epi)
plotTrace(mcmc.epi1$trace)
}
\seealso{
burnAndThin
}
