% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotTraj}
\alias{plotTraj}
\title{Plot one or more model trajectories}
\usage{
plotTraj(traj, state.variables = NULL, alpha = 1, plot = TRUE)
}
\arguments{
\item{traj}{data.frame, output of \code{fitmodel$simulate.model} or \code{simulateModelReplicates}.}

\item{state.variables}{subset of state.variables to plot. If \code{NULL} (default) all state variables are plotted.}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\description{
This function uses faceting to plot all trajectories in a data
frame. Convenient to visualise results of several simulations, or
data.
}
\seealso{
simulateModelReplicates
}

