% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{pmcmcSeit4lInfoPrior128}
\alias{pmcmcSeit4lInfoPrior128}
\title{Markov-chain Monte Carlo (MCMC) outputs of a fit of the stochastic SEIT4L
model to the influenza outbreak on Tristan da Cunha}
\format{
A list with of length 8 with each element containing three elements
  describing the MCMC outputs
}
\description{
A dataset containing the MCMC outputs of a fit of the `seit4lStoch` model to
the `fluTdc1971` data set with informative priors (8 chains of 3000
iterations each), using a particle filter with 128 particles to estimate the
likelihood at each MCMC setup. This differs from the `pMcmcSeitlInfoPrior8`
dataset in the number of particles.
}
\details{
\itemize{
  \item \code{trace} data frame containing the MCMC trace
  \item \code{acceptanceRate} a single value denoting the acceptance rate
  \item \code{covmatEmprical} the covariance matrix of parameter samples
}
}
\keyword{internal}
