% Generated by roxygen2 (4.0.1): do not edit by hand
\name{posteriorDensity}
\alias{posteriorDensity}
\title{Posterior distribution for a fitmodel}
\usage{
posteriorDensity(fitmodel, theta, state.init, data, margLogLike, ...)
}
\arguments{
\item{margLogLike}{\R-function to compute the marginal log-likelihood of \code{theta}.}

\item{...}{further arguments to be passed to \code{margLogLike}}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{state.init}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{time}, whereas the name of the data column should match the one used in the function \code{fitmodel$logLikePoint}.}
}
\value{
a list of two elements
\itemize{
	\item \code{log.density} numeric, logged value of the posterior density evaluated at \code{theta}
	\item \code{trace} named vector with trace information (theta, log.prior, marg.log.like, log.posterior)
}
}
\description{
This function evaluates the posterior distribution at \code{theta} and returns the result in a suitable format for \code{\link{mcmcMH}}.
}
\seealso{
\code{\link{margLogLikeDeter}}, \code{\link{margLogLikeSto}}
}

