% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{printNamedVector}
\alias{printNamedVector}
\title{Print named vector}
\usage{
printNamedVector(x, fmt = "\%.2f", sep = " | ")
}
\arguments{
\item{x}{named vector}

\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}
}
\description{
Print named vector with format specified by \code{fmt} (2 decimal places by
default).
}
\seealso{
\code{\link[base]{sprintf}}
}
\keyword{internal}
