% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLikelihood.r
\name{rTrajObs}
\alias{rTrajObs}
\title{Generate an observation trajectory for a fitmodel}
\usage{
rTrajObs(fitmodel, theta, init.state, times)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{init.state}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{times}{the times at which to generate observations}
}
\value{
numeric value of the log-likelihood
}
\description{
This function simulates a model defined in a \code{\link{fitmodel}}
object and generates observations at each time point. It returns
the trajectory with an additions \code{obs} column.
}
\seealso{
\code{\link{dTrajObs}}
}

