% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEIT2L_createModelTdC}
\alias{SEIT2L_createModelTdC}
\title{Create the SEIT2L model as a fitmodel object}
\usage{
SEIT2L_createModelTdC(simulate = c("deterministic", "stochastic"))
}
\arguments{
\item{simulate}{character. Either \code{"deterministic"} or \code{"stochastic"}.}
}
\value{
a fitmodel object
}
\description{
This function returns a fitmodel object contaning the (either deterministic or stochastic) SEIT2L model, to be fitted to the \code{\link{FluTdC1971}} dataset.
}

