% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{SEIT4L_stoch}
\alias{SEIT4L_stoch}
\title{The stochastic SEIT4L model with constant population size}
\format{A \code{\link{fitmodel}} object, that is a list with the following elements:}
\description{
The stochastic SEIT4L model with constant population size, uniform prior and Poisson observation with reporting rate.
}
\details{
\itemize{
  \item \code{name} character.
  \item \code{state.names} character vector.
  \item \code{theta.names} character vector.
  \item \code{simulate} \R-function.
  \item \code{rPointObs} \R-function.
  \item \code{dprior} \R-function.
  \item \code{dPointObs} \R-function.
}

Look at the documentation of \code{\link{fitmodel}} for more details about each of these elements.
You can look at the code of the \R-functions by typing \code{SEIT4L_stoch$simulate} for instance. There are some comments included.
}

