% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_createFitmodel}
\alias{SEITL_createFitmodel}
\title{Create the SEITL model as a fitmodel object}
\usage{
SEITL_createFitmodel(simulate = c("deterministic", "stochastic"))
}
\arguments{
\item{simulate}{character. Either \code{"deterministic"} or \code{"stochastic"}.}
}
\value{
a fitmodel object
}
\description{
This function returns a fitmodel object contaning the (either deterministic or stochastic) SEITL model, to be fitted to the \code{\link{FluTdC1971}} dataset.
}

