% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_genObsPoint}
\alias{SEITL_genObsPoint}
\title{Generate an observed incidence time series}
\usage{
SEITL_genObsPoint(model.point, theta)
}
\arguments{
\item{simu.traj}{data frame returned by \code{\link{SEITL_simulateDeterministic}} or \code{\link{SEITL_simulateStochastic}}}

\item{theta}{named numeric vector. Parameter values. Must include the following parameters: "R0", "LP", "IP", "alpha", "TIP" and "rho"}
}
\value{
the \code{simu.traj} data.frame with an additional variable: "observation".
}
\description{
Generate a daily incidence time serie under a Poisson observation process.
}
\seealso{
\code{\link{SEITL_simulateDeterministic}}, \code{\link{SEITL_simulateStochastic}}
}

