% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_logPrior}
\alias{SEITL_logPrior}
\title{Log-prior for SEITL model}
\usage{
SEITL_logPrior(theta)
}
\arguments{
\item{theta}{named numeric vector. Parameter values. Must include the following parameters: "R0", "LP", "IP", "alpha", "TIP" and "rho"}
}
\value{
the value of the log-prior.
}
\description{
Evaluate the log of the prior density distribution of the parameter values \code{theta}.
}
\seealso{
fitmodel, compositeLogPrior
}

