% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_simulateStochastic}
\alias{SEITL_simulateStochastic}
\title{Stochastic simulation of SEITL model}
\usage{
SEITL_simulateStochastic(theta, state.init, times)
}
\arguments{
\item{theta}{named numeric vector. Parameter values. Must include the following parameters: "R0", "D.lat", "D.inf", "alpha", "D.imm" and "rho"}

\item{state.init}{named numeric vector. Initial state of the SEITL model. Must include the following states: "S", "E", "I", "T", "L" and "Inc".}

\item{times}{numeric vector. Time sequence for which state of the model is wanted; the first value of times must be the initial time, i.e. the time of \code{state.init}.}
}
\value{
a \code{data.fame} containing the simulated trajectories that is the values of the state variables (1 per column) at each observation time (1 per row). The first column is \code{time}.
}
\description{
Simulate realisation of the stochastic version of the SEITL model using the \code{\link{simulateModelStochastic}} function.
}
\note{
The incidence \code{state.init[["Inc"]]} will be set to 0 at the start of the simulation and computed for each time interval of the vector \code{times}.
For instance, if \code{times} contains daily time steps, then the returned data frame will contain daily incidence.
}
\seealso{
\code{\link{fitmodel}}, \code{\link{simulateModelStochastic}}
}

