% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{seitlStoch}
\alias{seitlStoch}
\title{The stochastic SEITL model with constant population size}
\format{
A \code{\link{fitmodel}} object, that is a list with the following
  elements:
}
\description{
The stochastic SEITL model with constant population size, uniform prior and
Poisson observation with reporting rate.
}
\details{
\itemize{
  \item \code{name} character.
  \item \code{stateNames} character vector.
  \item \code{thetaNames} character vector.
  \item \code{simulate} \R-function.
  \item \code{rPointObs} \R-function.
  \item \code{dprior} \R-function.
  \item \code{dPointObs} \R-function.
}

Look at the documentation of \code{\link{fitmodel}} for more details about
each of these elements.
You can look at the code of the \R-functions by typing
\code{seitlStoch$simulate} for instance. There are some comments included.
}
\keyword{internal}
