% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{simulateModelReplicates}
\alias{simulateModelReplicates}
\title{Simulate several replicate of the model}
\usage{
simulateModelReplicates(
  fitmodel,
  theta,
  initState,
  times,
  n,
  observation = FALSE
)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should
match \code{fitmodel$thetaNames}.}

\item{initState}{named numeric vector. Initial values of the state
variables. Names should match \code{fitmodel$stateNames}.}

\item{times}{vector of times at which you want to observe the states of the
model.}

\item{n}{number of replicated simulations.}

\item{observation}{logical, if \code{TRUE} simulated observation are
generated by \code{rTrajObs}.}
}
\value{
a data.frame of dimension
  \code{[nxlength(times)]x[length(initState)+2]} with column names equal to
  \code{c("replicate","time",names(initState))}.
}
\description{
Simulate several replicate of a fitmodel using its function simulate
}
