% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{sir2popDeter}
\alias{sir2popDeter}
\title{The deterministic SIR model with two populations.}
\format{
A \code{\link{fitmodel}} object, that is a list with the following
  elements:
}
\description{
The deterministic SIR model with two populations (young/old) and constant
population size, uniform prior and Poisson observation.
}
\details{
\itemize{
  \item \code{name} character.
  \item \code{stateNames} character vector.
  \item \code{thetaNames} character vector.
  \item \code{simulate} \R-function.
  \item \code{rPointObs} \R-function.
  \item \code{dprior} \R-function.
  \item \code{dPointObs} \R-function.
}

Look at the documentation of \code{\link{fitmodel}} for more details about
each of these elements.
You can look at the code of the \R-functions by typing
\code{seitlDeter$simulate} for instance. There are some comments included.
}
