% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitR-package.r
\name{SIR_stoch}
\alias{SIR_stoch}
\title{A simple stochastic SIR model with constant population size}
\format{A \code{\link{fitmodel}} object, that is a list with the following elements:}
\description{
A simple stochastic SIR model with constant population size, uniform prior and Poisson observation.
}
\details{
\itemize{
  \item \code{name} character.
  \item \code{state.names} character vector.
  \item \code{theta.names} character vector.
  \item \code{simulate} \R-function.
  \item \code{genObsPoint} \R-function.
  \item \code{logPrior} \R-function.
  \item \code{pointLogLike} \R-function.
}

Look at the documentation of \code{\link{fitmodel}} for more details about each of these elements.
You can look at the code of the \R-functions by typing \code{SIR_stoch$simulate} for instance. There are some comments included.
}

