% Generated by roxygen2 (4.0.1): do not edit by hand
\name{targetPosterior}
\alias{targetPosterior}
\title{Target posterior distribution for a fitmodel}
\usage{
targetPosterior(theta, logPrior, logPrior.args = list(),
  marginal.logLikePoint, marginal.logLikePoint.args = list())
}
\arguments{
\item{theta}{named vector of estimated theta}

\item{logPrior}{\R-function to compute the log-prior of \code{theta}, as returned by \code{\link{fitmodel}}}

\item{logPrior.args}{list of arguments passed to \code{logPrior}}

\item{marginal.logLikePoint}{\R-function to compute the marginal log-likelihood of \code{theta}}

\item{marginal.logLikePoint.args}{list of arguments passed to \code{marginal.logLikePoint}}
}
\value{
a list of two elements
\itemize{
	\item \code{log.dist} numeric, logged value of the posterior distribution evaluated at \code{theta}
	\item \code{trace} named vector with trace information (theta, logPrior, marginal.logLikePoint, log.posterior)
}
}
\description{
This function evaluates the posterior distribution at \code{theta} and returns the result in a suitable format for \code{\link{mcmcMH}}.
}

