% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.fitr}
\alias{write.fitr}
\title{write function for class "fitr"}
\usage{
write.fitr(fitr, basefile = paste0(format(Sys.time(), "\%Y\%m\%d"), "fitR"),
  description = FALSE, sep = ";")
}
\arguments{
\item{object}{of class "fitr"}

\item{basefile}{base file name (not including .csv)}

\item{description}{data.frame containing description of your curves, must contain ID column with unique ID for each curve.}

\item{...}{arguments passed to \code{\link{write.csv}}}
}
\description{
writes contents of fitr object in a 3 .csv files, *_summary.csv, *_fits.csv and *_data.csv
}
\keyword{fitr}

