% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{burnAndThin}
\alias{burnAndThin}
\title{Burn and thin MCMC chain}
\usage{
burnAndThin(trace, burn = 0, thin = 0)
}
\arguments{
\item{trace}{either a \code{data.frame} or a \code{list} of \code{data.frame}
with all variables in column, as outputed by \code{\link{mcmcMh}}. Accept
also an \code{mcmc} or \code{mcmc.list} object.}

\item{burn}{proportion of the chain to burn.}

\item{thin}{number of samples to discard per sample that is being kept}
}
\value{
an object with the same format as \code{trace} (\code{data.frame} or
  \code{list} of \code{data.frame} or \code{mcmc} object or \code{mcmc.list}
  object)
}
\description{
Return a burned and thined trace of the chain.
}
\examples{
data(mcmcEpi)
burnAndThin(mcmcEpi1, burn = 100, thin = 4)
}
