% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeDIC}
\alias{computeDIC}
\title{Compute the DIC}
\usage{
computeDIC(trace, fitmodel, init.state, data, margLogLike = trajLogLike, ...)
}
\arguments{
\item{trace}{either a \code{data.frame} or \code{mcmc} object. Must contain one column with the posterior \code{log.likelihood}.}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{init.state}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{"time"} and the observation column must be named \code{"obs"}.}

\item{margLogLike}{\R-function to compute the marginal log-likelihood of \code{theta}.}

\item{...}{further arguments to be passed to \code{margLogLike}}
}
\value{
a list of 5 elements:
\itemize{
    \item \code{DIC} value of the DIC
    \item \code{theta_bar} mean posterior of theta
    \item \code{log_like_theta_bar} log-likelihood of \code{theta_bar}
    \item \code{D_theta_bar} deviance of \code{theta_bar}
    \item \code{p_D} effective number of parameters
}
}
\description{
This function computes the Deviance Information Criterion (DIC) of a \code{\link{fitmodel}} from a MCMC sample.
}

