% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitclasses.r
\name{fitmodel}
\alias{fitmodel}
\title{Constructor of fitmodel object}
\usage{
fitmodel(
  name = NULL,
  stateNames = NULL,
  thetaNames = NULL,
  simulate = NULL,
  rPointObs = NULL,
  dPrior = NULL,
  dPointObs = NULL
)
}
\arguments{
\item{name}{character. Name of the model (required).}

\item{stateNames}{character vector. Names of the state variables i.e.
\code{c("S","I","R")} (required).}

\item{thetaNames}{character vector. Names of the parameters i.e.
\code{c("R_0","infectiousPeriod")} (required).}

\item{simulate}{\R-function to simulate forward the model (required). This
  function takes 3 arguments:
\itemize{
\item \code{theta} named numeric vector. Values of the parameters. Names
  should match \code{thetaNames}.
\item \code{initState} named numeric vector. Initial values of the state
  variables. Names should match \code{stateNames}.
\item \code{times} numeric vector. Time sequence for which the state of the
  model is wanted; the first value of times must be the initial time, i.e.
  the time of \code{initState}.
}
and returns a \code{data.fame} containing the simulated trajectories that is
  the values of the state variables (1 per column) at each observation time
  (1 per row). The first column is \code{time}.}

\item{rPointObs}{\R-function that generates a (randomly sampled) observation
  point from a model point, using an observation model (optional). It thus
  acts as an inverse of \code{dPointObs} (see below). This function takes 2
  arguments
\itemize{
\item \code{modelPoint} named numeric vector. State of the model at a given
  point in time.
\item \code{theta} named numeric vector. Values of the parameters. Names
  should match \code{thetaNames}.
}
and returns an observation point}

\item{dPrior}{\R-function that evaluates the prior density of the parameters
  at a given \code{theta} (optional). The function should take 2 arguments:
\itemize{
    \item \code{theta} named numeric vector. Values of the parameters. Names
  should match \code{thetaNames}.
    \item \code{log} boolean. determines whether the logarithm of the prior
  density should be returned.
}
and returns the (logged, if requested) value of the prior density
  distribution.}

\item{dPointObs}{\R-function that evaluates the likelihood of one data point
  given the state of the model at the same time point. This function takes 4
  arguments:
\itemize{
\item \code{dataPoint} named numeric vector. Observation time and observed
  data point.
\item \code{modelPoint} named numeric vector containing the state of the
  model at the observation time point.
\item \code{theta} named numeric vector. Parameter values. Useful since
  parameters are usually needed to compute the likelihood (i.e. reporting
  rate).
\item \code{log} boolean. determines whether the logarithm of the likelihood
  should be returned.
}
and returns the (log-)likelihood. (optional)}
}
\value{
a \code{fitmodel} object that is a \code{list} of 7 elements:
\itemize{
	\item \code{name} character, name of the model
	\item \code{stateNames} vector, names of the state variables.
	\item \code{thetaNames} vector, names of the parameters.
	\item \code{simulate} \R-function to simulate forward the model; usage:
  \code{simulate(theta,initState,times)}.
	\item \code{rPointObs} \R-function to generate simulated observations;
  usage: \code{rPointObs(modelPoint, theta)}.
	\item \code{dPrior} \R-function to evaluate the log-prior of the parameter
  values; usage: \code{dPrior(theta)}.
	\item \code{dPointObs} \R-function to evaluate the log-likelihood of one
  data point; usage: \code{dPointObs(dataPoint, modelPoint, theta, log)}.
}
}
\description{
A \code{fitmodel} object is a \code{list} that stores some variables and
functions that will be useful to simulate and fit your model during the
course.
}
\seealso{
\code{\link{testFitmodel}}
}
