% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fitr}
\alias{fitr}
\alias{fitr-package}
\title{fitr.
A package containing tools for growth rate fitting of many growthcurves.}
\description{
\code{\link{gcfit}}
\code{\link{pickfit}}
\code{\link{plot_fitr}}
\code{\link{d_gcfit}}
\code{\link{print.fitr}}
\code{\link{plot.fitr}}
}

