% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR-package.r, R/fitR-class.R
\docType{package}
\name{fitR}
\alias{fitR}
\alias{fitR-package}
\alias{getfitR}
\alias{getfitR,fitR-method}
\title{fitR}
\usage{
getfitR(object = "fitR")

\S4method{getfitR}{fitR}(object = "fitR")
}
\description{
The fitR package calculates the fit of statistical models in relation to their 
predictive accuracy.
}
\examples{


}
\author{
David R. Miller: \email{drmiller@wustl.edu}
}
