% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitstat.R
\name{fitstat}
\alias{fitstat}
\title{Calculate Fit Statistics}
\usage{
fitstat(y = NULL, P, r = NULL, fits = c("RMSE", "MAD", "RMSLE", "MAPE",
  "MEAPE"), ...)
}
\arguments{
\item{y}{A vector of test data values}

\item{P}{A matrix of predicted values}

\item{r}{A vector of naive estimates}

\item{fits}{A vector of capitalized fit statistic abbreviations as strings. 
Possible values include RMSE, MAD, RMSLE, MAPE, MEAPE, and MRAE (for use with
r)}
}
\value{
An matrix of fit statistic values, where rows are models and columns 
are fit statistics
}
\description{
Finds fit statistics given a matrix of predicted values and either a vector
of test data or a vector of naive predictions, or both
}
\author{
Ryden W. Butler
}
