% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pickfit}
\alias{pickfit}
\title{pick best fits}
\usage{
pickfit(fits, min_numP, RsqCutoff = 0.95)
}
\arguments{
\item{fits}{\code{data.frame} containing fits generated by \code{\link{lfit}}}

\item{RsqCutoff}{Only fits with an adjusted R squared (see \code{\link{summary.lm}}) higher than the cutoff are considered. Defaults to 0.95.}
}
\description{
pick the 'most valid' fit. This will be the fit with the highest mumax and an adjusted R squared above \code{RsqCutoff}
}
\section{Output}{

   \code{data.frame} with all fits from lfit
}
\keyword{fitr,}
\keyword{growthcurve}

