% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lFitFunctions.R
\name{pickfit}
\alias{pickfit}
\title{pick best fits}
\usage{
pickfit(fits, min_numP, RsqCutoff = 0.95, growthCheck = "none")
}
\arguments{
\item{fits}{\code{data.frame} containing fits generated by \code{\link{gcfit}}}

\item{min_numP}{}

\item{RsqCutoff}{Only fits with an adjusted R squared (see \code{\link{summary.lm}}) higher than the cutoff are considered. Defaults to 0.95.}
}
\description{
pick the 'most valid' fit. This will be the fit with the highest mumax and an adjusted R squared above \code{RsqCutoff}
}
\section{Output}{

   \code{data.frame} best fits.
}
\keyword{fitr,}
\keyword{growthcurve}

