% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrClassFunctions.R
\name{plot.fitr}
\alias{plot.fitr}
\title{plot function for class "fitr"}
\usage{
\method{plot}{fitr}(fitr, interactive = TRUE, select = FALSE, sample_size = 5, save = FALSE)
}
\arguments{
\item{object}{of class "fitr"}
}
\description{
plots data and selected 'best' fits, a wrapper for \code{\link{plot_fitr}}
}
\keyword{fitr}
