% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.fitr}
\alias{plot.fitr}
\title{plot function for class "fitr"}
\usage{
\method{plot}{fitr}(fitr, interactive = TRUE, select = FALSE)
}
\arguments{
\item{object}{of class "fitr"}
}
\description{
plots data and best fits
}
\section{Output}{

   output description
}
\keyword{keyword}

