% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotEssBurn}
\alias{plotEssBurn}
\title{Plot Effective Sample Size (ESS) against burn-in}
\usage{
plotEssBurn(
  trace,
  longestBurnIn = ifelse(is.data.frame(trace) | is.matrix(trace) | is.mcmc(trace),
    nrow(trace), nrow(trace[[1]]))/2,
  stepSize = round(longestBurnIn/50)
)
}
\arguments{
\item{trace}{either a \code{data.frame} or a \code{list} of
\code{data.frame} with all variables in column, as outputed by
\code{\link{mcmcMh}}. Accept also \code{mcmc} or \code{mcmc.list} objects.}

\item{longestBurnIn}{The longest burn-in to test. Defaults to half the
length of the trace}

\item{stepSize}{The size of the steps of burn-in to test. Defaults to
1/50th of \code{longest.burn.in}}
}
\value{
a plot of the ESS against burn.in
}
\description{
Takes an mcmc trace and tests the ESS at different values of burn-in
}
\examples{
data(mcmcEpi)
plotEssBurn(mcmcEpi1)
}
