% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotfit}
\alias{plotfit}
\title{plot fits generated with lfit and selected with pickfit}
\usage{
plotfit(bestfit, fits, data, od_name, time_name, select = FALSE,
  interactive = TRUE)
}
\arguments{
\item{bestfit}{\code{data.frame} with best fit generated by \code{\link{pickfit}}}

\item{fits}{\code{data.frame} containing fits generated by \code{\link{lfit}}}

\item{data}{long-form data frame with growth data}

\item{od_name}{name of column containing OD values}

\item{time_name}{name of column containing times (in units after start of experiment)}

\item{select}{a vector containing IDs of a selection of fits to plot. Defaults to FALSE, showing all fits. IDs not found in \code{data} are quietly ignored.}

\item{interactive}{boolean; Should every plot be shown? Advance to next plot by clicking in plot area.}
}
\description{
plot all fits, or only a selection using the \code{select} argument.
}
\section{Output}{

   none
}
\keyword{growthcurve}

