% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPosteriorDensity}
\alias{plotPosteriorDensity}
\title{Plot MCMC posterior densities}
\usage{
plotPosteriorDensity(trace)
}
\arguments{
\item{trace}{either a \code{data.frame} or a \code{list} of \code{data.frame} with all variables in column, as outputed by \code{\link{mcmcMH}}. Accept also an \code{mcmc} or \code{mcmc.list} object.}
}
\description{
Plot the posterior density.
}
\seealso{
burnAndThin
}

