% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPosteriorFit}
\alias{plotPosteriorFit}
\title{Plot MCMC posterior fit}
\usage{
plotPosteriorFit(trace, fitmodel, state.init, data,
  posterior.summary = c("sample", "median", "mean", "max"), summary = TRUE,
  sample.size = 100, alpha = min(1, 10/sample.size), plot = TRUE,
  all.vars = FALSE)
}
\arguments{
\item{posterior.summary}{character. Set to \code{"sample"} to plot trajectories from a sample of the posterior (default). Set to \code{"median"}, \code{"mean"} or \code{"max"} to plot trajectories corresponding to the median, mean and maximum of the posterior density.}

\item{summary}{logical, if \code{TRUE} trajectories are summarised by their mean, median, 50\% and 95\% quantile distributions. Otheriwse, the trajectories are ploted.}

\item{sample.size}{number of theta sampled from posterior distribution (if \code{posterior.summary=="sample"}). Otherwise, number of replicated simulations.}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{state.init}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{"time"} and the observation column must be named \code{"obs"}.}

\item{trace}{a \code{data.frame} with one column per estimated parameter, as returned by \code{\link{burnAndThin}}}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\value{
If \code{plot==FALSE}, a list of 2 elements is returned:
\itemize{
   \item \code{posterior.traj} a \code{data.frame} with the trajectories (and observations) sampled from the posterior distribution.
   \item \code{plot} the plot of the fit displayed.
}
}
\description{
Plot posterior distribution of observation generated under model's posterior parameter distribution.
}

