% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotTraj}
\alias{plotTraj}
\title{Plot model trajectories}
\usage{
plotTraj(traj = NULL, state.names = NULL, data = NULL, summary = TRUE,
  alpha = 1, plot = TRUE)
}
\arguments{
\item{traj}{data.frame, output of \code{fitmodel$simulate} or \code{simulateModelReplicates}.}

\item{state.names}{character vector. Names of the state variables to plot. Names must match \code{fitmodel$state.names}. If \code{NULL} (default) all state variables are plotted.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{time}, whereas the name of the data column should match one of \code{fitmodel$state.names}.}

\item{summary}{logical. If \code{TRUE}, the mean, median as well as the 50th and 95th percentile of the trajectories are plotted (default). If \code{FALSE}, all individual trajectories are plotted (transparency can be set with \code{alpha}).}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\description{
This function use faceting to plot all trajectories in a data frame. Convenient to see results of several simulations, or data. Also, if \code{data} is present, then an additional plot is displayed with data and potentially observation generated.
}
\seealso{
\code{\link{simulateModelReplicates}}
}

