% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrClassFunctions.R
\name{print.fitr}
\alias{print.fitr}
\title{print function for class "fitr"}
\usage{
\method{print}{fitr}(fitr)
}
\arguments{
\item{object}{of class "fitr"}
}
\description{
prints description of data and selected 'best' fits
}
\keyword{fitr}

