% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_genObsPoint}
\alias{SEITL_genObsPoint}
\title{Generate an observed incidence}
\usage{
SEITL_genObsPoint(model.point, theta)
}
\arguments{
\item{model.point}{named numeric vector. State of the model at a given point in time.}

\item{theta}{named numeric vector. Values of the parameters.}
}
\value{
the \code{simu.traj} data.frame with an additional variable: "observation".
}
\description{
Generate an observed incidence under a Poisson observation process.
}
\seealso{
\code{\link{SEITL_simulateDeterministic}}, \code{\link{SEITL_simulateStochastic}}
}

