% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulateFinalStateAtExtinction}
\alias{simulateFinalStateAtExtinction}
\title{Simulate model until extinction}
\usage{
simulateFinalStateAtExtinction(fitmodel, theta, init.state, extinct = NULL,
  time.init = 0, time.step = 100, n = 100, observation = FALSE,
  n.cores = 1)
}
\arguments{
\item{extinct}{character vetor. Simulations stop when all these state are extinct.}

\item{time.init}{numeric. Start time of simulation.}

\item{time.step}{numeric. Time step at which extinction is checked}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{init.state}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{n}{number of replicated simulations.}

\item{observation}{logical, if \code{TRUE} simulated observation are generated by \code{genObsTraj}.}

\item{n.cores}{number of cores on which propogation of the particles is parallelised. By default no parallelisation (\code{n.cores=1}). If \code{NULL}, set to the value returned by \code{\link[parallel]{detectCores}}.}
}
\description{
Return final state at extinction
}

