% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{simulateFinalStateAtExtinction}
\alias{simulateFinalStateAtExtinction}
\title{Simulate model until extinction}
\usage{
simulateFinalStateAtExtinction(
  fitmodel,
  theta,
  initState,
  extinct = NULL,
  timeInit = 0,
  timeStep = 100,
  n = 100,
  observation = FALSE,
  nCores = 1
)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should
match \code{fitmodel$thetaNames}.}

\item{initState}{named numeric vector. Initial values of the state
variables. Names should match \code{fitmodel$stateNames}.}

\item{extinct}{character vetor. Simulations stop when all these state are
extinct.}

\item{timeInit}{numeric. Start time of simulation.}

\item{timeStep}{numeric. Time step at which extinction is checked}

\item{n}{number of replicated simulations.}

\item{observation}{logical, if \code{TRUE} simulated observation are
generated by \code{rTrajObs}.}
}
\description{
Return final state at extinction
}
\keyword{internal}
