% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulateModelReplicates}
\alias{simulateModelReplicates}
\title{Simulate several replicate of the model}
\usage{
simulateModelReplicates(fitmodel, theta, state.init, times, n,
  observation = FALSE)
}
\arguments{
\item{times}{vector of times at which you want to observe the states of the model.}

\item{n}{number of replicated simulations.}

\item{observation}{logical, if \code{TRUE} simulated observation are generated by \code{genObsTraj}.}

\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{state.init}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}
}
\value{
a data.frame of dimension \code{[nxlength(times)]x[length(state.init)+2]} with column names equal to \code{c("replicate","time",names(state.init))}.
}
\description{
Simulate several replicate of a fitmodel using its function simulate
}

