% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitclasses.r
\name{testFitmodel}
\alias{testFitmodel}
\title{Test a fitmodel}
\usage{
testFitmodel(fitmodel, theta, initState, data = NULL, verbose = TRUE)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should
match \code{fitmodel$thetaNames}.}

\item{initState}{named numeric vector. Initial values of the state
variables. Names should match \code{fitmodel$stateNames}.}

\item{data}{data frame. Observation times and observed data. The time column
must be named \code{"time"} and the observation column must be named
\code{"obs"}.}

\item{verbose}{if \code{TRUE}, print details of the test performed to check
validity of the arguments}
}
\description{
This function performs a series of checks on the \code{\link{fitmodel}}
provided by the user in order to make sure that it will be compatible both
with the functions coded during the course and the functions available in the
\code{fitR} package.
}
\seealso{
\code{\link{fitmodel}}
}
