% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trajLogLike}
\alias{trajLogLike}
\title{Log-likelihood of a trajectory for a deterministic model}
\usage{
trajLogLike(fitmodel, theta, init.state, data)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object}

\item{theta}{named numeric vector. Values of the parameters. Names should match \code{fitmodel$theta.names}.}

\item{init.state}{named numeric vector. Initial values of the state variables. Names should match \code{fitmodel$state.names}.}

\item{data}{data frame. Observation times and observed data. The time column must be named \code{"time"} and the observation column must be named \code{"obs"}.}
}
\value{
numeric value of the log-likelihood
}
\description{
Compute the trajectory log-likelihood of \code{theta} for a
deterministic model defined in a \code{\link{fitmodel}} object by
summing the point log-likelihoods.
}
\seealso{
\code{\link{genObsTraj}}
}

